Time: 7 ms (70.48%) | Memory: 21.4 MB (21.32%) - LeetSync

+++ 26-remove-duplicates-from-sorted-array/remove-duplicates-from-sorted-array.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int removeDuplicates(vector<int>& nums) {
-        long long i =0;
+        int i =0;
         for(int j =1;j<nums.size();j++){
             if(nums[j]!=nums[i]){
                 i=i+1;
